% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_bbox.R
\name{ggmap_bbox}
\alias{ggmap_bbox}
\title{Convert the bounding box of a ggmap object to correspond to a sf object w EPSG4326}
\usage{
ggmap_bbox(map)
}
\arguments{
\item{map}{a ggmap object, created with the getmap() function}
}
\value{
a ggmap object with adjusted CRS (to EPSG4326)
}
\description{
Occasionally, when adding a sf object to a ggmap, there is misalignment of the sf layer and the background map. This code adjusts the ggmap object to match the sf object.
Code from: https://stackoverflow.com/questions/47749078/how-to-put-a-geom-sf-produced-map-on-top-of-a-ggmap-produced-raster/50844502#50844502
}
