% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotekDownload.R
\name{lotekDownload}
\alias{lotekDownload}
\title{Download Lotek collar data from WebService using API}
\usage{
lotekDownload(Username, Password, StartDate, EndDate)
}
\arguments{
\item{Username}{Webservice username credential}

\item{Password}{Webservice password credential}

\item{StartDate}{beginning date (local MST, "YYYY-MM-DD") of download. Assumes from 12AM onwards and accounts for daylight savings}

\item{EndDate}{ending date (local MST, "YYYY-MM-DD" or Sys.Date()) of download. Data downloaded is inclusive of this date. Assumes up to 11:59PM and accounts for daylight savings}
}
\value{
A dataframe of collar fix locations occurring within the specified start and end dates, arranged by device SN and date. Output intended to mimic that of manually downloaded data from Webservice (one notable consequence is that it only outputs GMT, not local MST).
}
\description{
This function downloads Lotek collar data and outputs a dataframe.
}
