% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLines.R
\name{toLines}
\alias{toLines}
\title{Convert collar point locations to an sf line geometry object}
\usage{
toLines(x, indID, datetime, lat, long)
}
\arguments{
\item{x}{data.frame or sf point geometry object containing collar locations}

\item{indID}{name of the field identifying individual animals, in quotes}

\item{datetime}{name of the field identifying DateTime, in quotes}

\item{lat}{if data.frame, name of the field containing the latitude, in quotes}

\item{long}{if data.frame, name of the field containing the longitude, in quotes}
}
\value{
sf object containing line feature geometries by individual
}
\description{
This function converts a dataframe or sf point object containing collar locations for individual animals to an sf object containing line geometry features for each individual. Retains all fields & attributes from the input dataset.
}
\details{
Note:
\itemize{
\item Coordinates must be in WGS84 - Lat/Long
\item Assumes collar data has been "cleaned"
}
}
