% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotekDownload.R
\name{lotekDownload}
\alias{lotekDownload}
\title{Download Lotek collar data from WebService using API}
\usage{
lotekDownload(Username, Password, StartDate, EndDate)
}
\arguments{
\item{Username}{Webservice username credential}

\item{Password}{Webservice password credential}

\item{StartDate}{beginning date of download, assumes from 12AM onwards}

\item{EndDate}{Webservice username credential, assumes up to 11:59PM}
}
\value{
A dataframe of collar locations within the provided start and end dates, arranged by device SN and date
}
\description{
This function downloads Lotek collar data and outputs a dataframe.
}
