% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLines.R
\name{toLines}
\alias{toLines}
\title{Convert collar point locations to an sf line geometry object}
\usage{
toLines(x, indID, datetime, lat, long)
}
\arguments{
\item{x}{object of data.frame or sf point geometry object containing collar locations}

\item{indID}{name of the field identifying individual animals}

\item{datetime}{name of the field identifying DateTime}

\item{lat}{if data.frame, name of the field containing the latitude}

\item{long}{if data.frame, name of the field containing the longitude}
}
\value{
A sf line geometry layer by individual
}
\description{
This function converts a dataframe or sf point object containing collar locations for individual animals to an sf object containing line geometry features for each individual.
Note:
\itemize{
\item Coordinates must be in WGS84 - Lat/Longs
\item Assumes collar data has been "cleaned"
}
}
